requirements.txt
sh
idna
wrapt
PyJWT
certifi
inflect
chardet
urllib3
requests
PyGithub
Deprecated
configparser
